Security

+++ app/src/main/java/com/muzima/controller/FormController.java
@@ -217,13 +217,6 @@ public class FormController {
         } catch (IOException e) {
             throw new FormFetchException(e);
         }
-        for (Form form : allForms) {
-            for (Tag tag : form.getTags()) {
-                if (!allTags.contains(tag)) {
-                    allTags.add(tag);
-                }
-            }
-        }
         return allTags;
     }
 
@@ -1080,11 +1073,13 @@ public class FormController {
                 if(parseAsFormData || !formData.getDiscriminator().equals(Constants.FORM_JSON_DISCRIMINATOR_INDIVIDUAL_OBS)){
                     if (isCompleteFormData(formData) || isArchivedFormData(formData)) {
                         List<Encounter> encounters = encounterService.getEncountersByFormDataUuid(formData.getUuid());
-                        for (Encounter encounter : encounters) {
-                            List<Observation> observations = observationService.getObservationsByEncounter(encounter.getEncounterUuid());
-                            observationService.deleteObservations(observations);
+                        if(encounters != null) {
+                            for (Encounter encounter : encounters) {
+                                List<Observation> observations = observationService.getObservationsByEncounter(encounter.getEncounterUuid());
+                                observationService.deleteObservations(observations);
+                            }
+                            encounterService.deleteEncounters(encounters);
                         }
-                        encounterService.deleteEncounters(encounters);
                     }
                     formService.deleteFormData(formData);
                 }
